move folder: test -> tests
