+++ tests/test_modif_1_4.py
@@ -4,7 +4,7 @@ import pytest
 import laspy
 from laspy import LazBackend
 from laspy.errors import LaspyError
-from test.test_common import test1_4_las, write_then_read_again
+from tests.test_common import test1_4_las, write_then_read_again
 
 
 @pytest.fixture(scope="session")
