+++ tests/test_reading_1_2.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 import laspy
-from test.test_common import (
+from tests.test_common import (
     simple_las,
     simple_laz,
 )
