+++ tests/vegetation_1_3.las
